# docs/docs_swagger/swagger.yaml
basePath: /api
definitions:
...
  field.FileType:
    enum:
    - pdf
    type: string
    x-enum-varnames:
    - FileTypePdf
  field.NoticeType:
    enum:
    - urgent
    - normal
    - building
    - government
    type: string
    x-enum-varnames:
    - NoticeTypeUrgent
    - NoticeTypeNormal
    - NoticeTypeBuilding
    - NoticeTypeGovernment
  field.Status:
    enum:
    - pending
    - active
    - inactive
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusActive
    - StatusInactive
...
  http_base_controller.CreateNoticeRequest:
    properties:
      description:
        example: 关于召开2023年度业主大会的通知
        type: string
      endTime:
        example: "2023-06-30T23:59:59Z"
        type: string
      fileType:
        allOf:
        - $ref: '#/definitions/field.FileType'
        example: pdf
      isIsmartNotice:
        example: false
        type: boolean
      isPublic:
        example: true
        type: boolean
      path:
        example: /uploads/documents/owners_meeting.pdf
        type: string
      priority:
        example: 1
        type: integer
      startTime:
        example: "2023-06-01T00:00:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/field.Status'
        example: active
      title:
        example: 业主大会通知
        type: string
      type:
        allOf:
        - $ref: '#/definitions/field.NoticeType'
        example: normal
    required:
    - endTime
    - fileType
    - path
    - startTime
    - status
    - title
    - type
    type: object
...
host: localhost:10031
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 智能门禁管理系统API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /admin/notice:
    delete:
      consumes:
      - application/json
      description: 删除一个或多个通知
      parameters:
      - description: 通知ID列表
        in: body
        name: ids
        required: true
        schema:
          type: object
      - description: 通知ID数组
        in: body
        name: ids.ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除通知
      tags:
      - Notice
    get:
      consumes:
      - application/json
      description: 根据查询条件获取通知列表
      parameters:
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 通知类型
        in: query
        name: type
        type: string
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - default: 1
        description: 页码
        in: query
        name: pageNum
        type: integer
      - default: false
        description: 是否降序
        in: query
        name: desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 返回通知列表和分页信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取通知列表
      tags:
      - Notice
    post:
      consumes:
      - application/json
      description: 创建一个新的通知
      parameters:
      - description: 通知信息
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/http_base_controller.CreateNoticeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建的通知信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建通知
      tags:
      - Notice
    put:
      consumes:
      - application/json
      description: 更新通知信息
      parameters:
      - description: 通知更新信息
        in: body
        name: notice
        required: true
        schema:
          type: object
      - description: 通知ID
        in: formData
        name: id
        required: true
        type: integer
      - description: 通知标题
        in: formData
        name: title
        type: string
      - description: 通知描述
        in: formData
        name: description
        type: string
      - description: 通知类型
        in: formData
        name: type
        type: string
      - description: 状态
        in: formData
        name: status
        type: string
      - description: 开始时间
        in: formData
        name: startTime
        type: string
      - description: 结束时间
        in: formData
        name: endTime
        type: string
      - description: 是否公开
        in: formData
        name: isPublic
        type: boolean
      - description: 优先级
        in: formData
        name: priority
        type: integer
      - description: 文件路径
        in: formData
        name: path
        type: string
      - description: 文件类型
        in: formData
        name: fileType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新后的通知信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新通知
      tags:
      - Notice
...
  /admin/notice/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取通知详细信息
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回通知详细信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 通知不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取单个通知
      tags:
      - Notice
  /admin/notices:
    post:
      consumes:
      - application/json
      description: 批量创建多个通知
      parameters:
      - description: 通知信息数组
        in: body
        name: notices
        required: true
        schema:
          items:
            $ref: '#/definitions/http_base_controller.CreateNoticeRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建的通知信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量创建通知
      tags:
      - Notice
...
securityDefinitions:
  BearerAuth:
    description: 输入带有Bearer前缀的用户认证token
    in: header
    name: Authorization
    type: apiKey
  DeviceAuth:
    description: 输入带有Bearer前缀的设备认证token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"